Update patch set 1
Patch Set 1:

> Uploaded patch set 1.

That was an annoying problem for quite some time - because broken queries can't be serialized, they are passed as null-queries to Krill, resulting in cascading errors. This removes at least the Krill error. But I think we can merge the errors even further. Probably in Koral and Kalamar.

Patch-set: 1
Label: Code-Review=0