Update patch set 5
Patch Set 5:

The equivalent Docker-compose to this environment would be, btw.

  version: '3.8'
  services:
    kalamar:
      image: "korap/kalamar"
      ports:
        - "64543:64543"
      environment:
        KALAMAR_API: "http://kustvakt:8089/api/"
    kustvakt:
      image: "korap/kustvakt"
      expose:
        - "8089"
      volumes:
        - "${INDEX}:/kustvakt/index"

This file (in a "docker-compose.yml") can then be started with

  $ INDEX=[LOCALINDEX] docker-compose up

and Korap should be up and running. However - this would ignore for the moment all configurations relevant for Kalamar and Kustvakt to be used in a production environment. I haven't played enough with "configs" in docker-compose yet to come up with an elegant solution yet.

Patch-set: 5