Jersey 2: Create a custom annotation to register resource filters
There is no direct equivalent of the @ResourceFilters annotation in Jersey 2.
The best alternative to define a custom annotation and use a DynamicFeature provider
to interpret this annotation and register filters.

Why not define a annotation for each filter and use Jersey's @NameBinding mechanism?
@ResourceFilters works differently than @NameBinding. A @ResourceFilters annotation
on method overrides the class level annotation; In this case, all filters
defined on the class level are discarded. This behaviour can't be achieved
with @NameBinding.

References:
https://eclipse-ee4j.github.io/jersey.github.io/documentation/2.37/filters-and-interceptors.html#d0e10127
https://jakartaee.github.io/rest/apidocs/2.1.6/javax/ws/rs/container/DynamicFeature.html