Jersey 2: Fix curly braces in query parameters
Jersey parses URI query parameters in order to expand templates; A template
placeholder is an identifier enclosed in curly braces.
To be able to use curly braces in URI query parameters, an indirection is required.
The query parameter will hold a template placeholder which will expand to
the real value of the query parameter which can contain curly braces;
This value will not be expanded further.

References:
https://jakartaee.github.io/rest/apidocs/2.1.6/javax/ws/rs/client/WebTarget.html#queryParam(java.lang.String,java.lang.Object...)
https://jakartaee.github.io/rest/apidocs/2.1.6/javax/ws/rs/client/WebTarget.html#resolveTemplate(java.lang.String,java.lang.Object)